Implemented backend services enabling followers, real-time post notifications, and content feeds in Chatbook.

+++ src/main/java/in/krish/repo/PostRepo.java
@@ -8,16 +8,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface PostRepo extends JpaRepository<Post, Integer> {
+public interface PostRepo extends JpaRepository<Post, Long> {
 
     List<Post> findByUser(User user);
-
-//    @Query("SELECT new in.krish.binding.PostSummaryDto(" +
-//            "p.title, p.content, p.createdAt, COUNT(DISTINCT l), COUNT(DISTINCT c)) " +
-//            "FROM Post p " +
-//            "LEFT JOIN p.likes l " +
-//            "LEFT JOIN p.comments c " +
-//            "GROUP BY p.id, p.title, p.content, p.createdAt")
-//    List<PostSummaryDto> fetchAllPostSummaries();
 
 }
